Move all helm releated functions into util.go
